added jwt token auth and made hotel creation protected

+++ src/main/java/com/bookhotel/hotelmanagement/api/service/impl/HotelServiceImpl.java
@@ -4,10 +4,12 @@ import com.bookhotel.hotelmanagement.api.dto.HotelDto;
 import com.bookhotel.hotelmanagement.api.dto.RoomDto;
 import com.bookhotel.hotelmanagement.api.entity.Hotel;
 import com.bookhotel.hotelmanagement.api.entity.Room;
+import com.bookhotel.hotelmanagement.api.entity.User;
 import com.bookhotel.hotelmanagement.api.repository.HotelRepository;
+import com.bookhotel.hotelmanagement.api.repository.UserRepository;
 import com.bookhotel.hotelmanagement.api.service.HotelService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.web.authentication.Http403ForbiddenEntryPoint;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -16,10 +18,12 @@ import java.util.*;
 public class HotelServiceImpl implements HotelService {
 
     private final HotelRepository hotelRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public HotelServiceImpl(HotelRepository hotelRepository) {
+    public HotelServiceImpl(HotelRepository hotelRepository, UserRepository userRepository) {
         this.hotelRepository = hotelRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
@@ -35,10 +39,14 @@ public class HotelServiceImpl implements HotelService {
     @Override
     public Hotel create(HotelDto hotelDto) {
 
+        User user = userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElse(null);
+
         Hotel hotel = Hotel.builder()
                 .name(hotelDto.getName())
                 .description(hotelDto.getDescription())
                 .address(hotelDto.getAddress())
+                .owner(user)
                 .build();
 
         hotel.setRooms(roomsConverting(hotelDto.getRooms(), hotel));
@@ -56,9 +64,8 @@ public class HotelServiceImpl implements HotelService {
 
         Hotel existingHotel = findById(id);
 
-        if (!hotelDto.getOwner().equals(existingHotel.getOwner().getUsername())) {
-            return null;
-        }
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!existingHotel.getOwner().getUsername().equals(username)) return null; // Forbidden
 
         if (hotelDto.getName() != null) existingHotel.setName(hotelDto.getName());
         if (hotelDto.getDescription() != null) existingHotel.setDescription(hotelDto.getDescription());
@@ -70,6 +77,12 @@ public class HotelServiceImpl implements HotelService {
 
     @Override
     public void deleteById(Long id) {
+
+        Hotel hotel = findById(id);
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!hotel.getOwner().getUsername().equals(username)) return; // Forbidden
+
         hotelRepository.deleteById(id);
     }
